requirements.txt
pyqtgraph
python-rtmidi
numpy
